first draft of the directory importer
