+++ src/directory/api.js
@@ -1,0 +1,7 @@
+// @ts-nocheck
+
+/**
+ * This is file is shadowed by api.ts and only exists so JS could import it to
+ * refer to types without encountering runtime errors
+ * @see https://github.com/microsoft/TypeScript/issues/41825
+ */
